requirements.txt
streamlit==1.36.0
pypdf==4.2.0
langchain==0.2.11
langchain-community==0.2.10
langchain-text-splitters==0.2.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
torch==2.3.1
python-dotenv==1.0.1
google-generativeai==0.7.2